Shadify (#150)
Adds shadcn, bumps dependencies, overhaults lots of code.

+++ src/app/(drift)/admin/components/tables.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import Button from "@components/button"
+import { Button } from "@components/button"
 import { Spinner } from "@components/spinner"
 import { useToasts } from "@components/toasts"
 import { ServerPostWithFilesAndAuthor, UserWithPosts } from "@lib/server/prisma"
@@ -18,7 +18,7 @@ export function UserTable({
 		id: string
 		email: string | null
 		role: string | null
-		displayName: string | null
+		username: string | null
 	}[]
 }) {
 	const { setToast } = useToasts()
@@ -26,6 +26,14 @@ export function UserTable({
 
 	const deleteUser = async (id: string) => {
 		try {
+			const confirmed = confirm("Are you sure you want to delete this user?")
+			if (!confirmed) {
+				setToast({
+					message: "User not deleted",
+					type: "default"
+				})
+				return
+			}
 			const res = await fetchWithUser("/api/admin?action=delete-user", {
 				method: "DELETE",
 				headers: {
@@ -53,8 +61,8 @@ export function UserTable({
 	}
 
 	return (
-		<table className={styles.table}>
-			<thead>
+		<table className="w-full overflow-x-auto">
+			<thead className="text-left">
 				<tr>
 					<th>Name</th>
 					<th>Email</th>
@@ -73,14 +81,20 @@ export function UserTable({
 				) : null}
 				{users?.map((user) => (
 					<tr key={user.id}>
-						<td>{user.displayName ? user.displayName : "no name"}</td>
+						<td>{user.username ? user.username : "no name"}</td>
 						<td>{user.email}</td>
 						<td>{user.role}</td>
 						<td className={styles.id} title={user.id}>
 							{user.id}
 						</td>
 						<td>
-							<Button onClick={() => deleteUser(user.id)}>Delete</Button>
+							<Button
+								variant={"destructive"}
+								onClick={() => deleteUser(user.id)}
+								size={"sm"}
+							>
+								Delete
+							</Button>
 						</td>
 					</tr>
 				))}
@@ -90,7 +104,7 @@ export function UserTable({
 }
 
 export function PostTable({
-	posts
+	posts: initialPosts
 }: {
 	posts?: {
 		createdAt: string
@@ -100,15 +114,54 @@ export function PostTable({
 		visibility: string
 	}[]
 }) {
+	const [posts, setPosts] = useState<typeof initialPosts>(initialPosts)
+	const { setToast } = useToasts()
+	const deletePost = async (id: string) => {
+		try {
+			const confirmed = confirm("Are you sure you want to delete this post?")
+			if (!confirmed) {
+				setToast({
+					message: "Post not deleted",
+					type: "default"
+				})
+				return
+			}
+			const res = await fetchWithUser("/api/admin?action=delete-post", {
+				method: "DELETE",
+				headers: {
+					"Content-Type": "application/json"
+				},
+				body: JSON.stringify({
+					postId: id
+				})
+			})
+
+			if (res.status === 200) {
+				setToast({
+					message: "Post deleted",
+					type: "success"
+				})
+				setPosts(posts?.filter((post) => post.id !== id))
+			}
+		} catch (err) {
+			console.error(err)
+			setToast({
+				message: "Error deleting user",
+				type: "error"
+			})
+		}
+	}
+
 	return (
-		<table className={styles.table}>
-			<thead>
+		<table className="w-full overflow-x-auto">
+			<thead className="text-left">
 				<tr>
 					<th>Title</th>
 					<th>Author</th>
 					<th>Created</th>
 					<th>Visibility</th>
 					<th className={styles.id}>Post ID</th>
+					<th>Actions</th>
 				</tr>
 			</thead>
 			<tbody>
@@ -130,6 +183,15 @@ export function PostTable({
 						<td>{new Date(post.createdAt).toLocaleDateString()}</td>
 						<td>{post.visibility}</td>
 						<td>{post.id}</td>
+						<td>
+							<Button
+								variant={"destructive"}
+								size={"sm"}
+								onClick={() => deletePost(post.id)}
+							>
+								Delete
+							</Button>
+						</td>
 					</tr>
 				))}
 			</tbody>
